chore: filter active accounts and refactor @types
